v3.8 : Level 2 platforms physics features added && water droping with boxes

+++ src/particleSystem/FireParticleSystem.java
@@ -25,7 +25,7 @@ public class FireParticleSystem implements ParticleSystemFactory
     }
     
     @Override
-    public SpriteParticleSystem build(Engine engine, int fontX, int fontY)
+    public SpriteParticleSystem build(Engine engine, float fontX, float fontY)
     {
     	//X & Y for the particles to spawn at.
 	    final float particlesXSpawn = fontX;
